Vendor `gardener@1.81.0` and implement `ForceDelete` method for extensions (#652)
* Vendor `gardener@v1.81.0`

* Implement `ForceDelete` for controllers

* Adapt breaking changes

+++ pkg/controller/controlplane/add.go
@@ -63,7 +63,7 @@ func AddToManagerWithOptions(ctx context.Context, mgr manager.Manager, opts AddO
 	genericActuator, err := genericactuator.NewActuator(mgr, alicloud.Name, secretConfigsFunc, shootAccessSecretsFunc, nil, nil,
 		nil, controlPlaneChart, controlPlaneShootChart, controlPlaneShootCRDsChart, storageClassChart, nil,
 		NewValuesProvider(mgr, opts.CSI), extensionscontroller.ChartRendererFactoryFunc(util.NewChartRendererForShoot),
-		imagevector.ImageVector(), "", opts.ShootWebhookConfig, opts.WebhookServerNamespace, defaultServer.Options.Port)
+		imagevector.ImageVector(), "", opts.ShootWebhookConfig, opts.WebhookServerNamespace, int32(defaultServer.Options.Port))
 	if err != nil {
 		return err
 	}
